chatbot/requirements.txt
Flask>=3.1.1
numpy>=2.3.1
neo4j>=5.28.1
python-dotenv>=1.1.1
google-cloud-aiplatform>=1.101.0
google-genai
langchain>=0.3.26
dspy>=2.6.27
langgraph>=0.5.1
openai
langchain_google_genai
flask_cors
